Add pep_0_errors.py
